:tracktor: rename js_extension to js_extensions
