+++ test/test_js_extension.py
@@ -2,7 +2,7 @@ import os
 from mock import patch
 from nose.tools import eq_
 
-from pyecharts.js_extension import JsExtension
+from pyecharts.js_extensions import JsExtension
 
 
 def produce_test_js_extension():
